Task 4, Strategy finished

+++ lab2/lab2.task4/src/main/java/percentile/PercentileCalculator.java
@@ -4,6 +4,6 @@ import java.util.List;
 
 public interface PercentileCalculator {
 
-    Long calculate(Integer p, List<Long> elements);
+    Long calculate(Double p, List<Long> elements);
 
 }
